+++ lib/ember_cli/html_constraint.rb
@@ -1,9 +1,0 @@
-module EmberCli
-  class HtmlConstraint
-    def matches?(request)
-      matches = request.format.to_s =~ /html/ || -1
-
-      matches > -1
-    end
-  end
-end